firefly/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2